+++ tasks/add-preview/lambda_function.py
@@ -69,10 +69,10 @@ def lambda_handler(payload, context={}):
 
             # add preview to item
             
-            add_preview(item, item['assets'][asset]['href'], **config)
+            item = add_preview(item, item['assets'][asset]['href'], **config)
             if thumb:
                 # add thumbnail to item
-                add_thumbnail(item, item['assets']['preview']['href'])
+                item = add_thumbnail(item, item['assets']['preview']['href'])
 
             # put back original href
             item['assets'][asset]['href'] = href
@@ -102,7 +102,7 @@ def lambda_handler(payload, context={}):
 
 def add_thumbnail(item, filename, scale_percent=5):
     """ Add a thumbnail to item, generated from filename """
-    fnout = filename.replace('.tif', '.png')
+    fnout = filename.replace('_preview.tif', '_thumb.png')
     logging.info(f"Creating thumbnail {fnout} from {filename}")
     try:
         gdal.Translate(fnout, filename, format='PNG', widthPct=scale_percent, heightPct=scale_percent)